Move `box_cxcywh_to_xyxy` method from metrics to model
* Clarify which class is responsible for the output for more concise and simple code
* The `DETR` model will produce output as `xyxy` form now
